pyproject.toml
[tool.poetry]
name = "senior_pomidor2"
version = "0.1.0"
description = ""
authors = ["anton <sparco114@ya.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "3.2.19"
psycopg2 = "2.9.6"
djangorestframework = "^3.14.0"
django-filter = "23.2"
social-auth-app-django = "5.2.0"
coverage = "^7.2.7"
django-debug-toolbar = "^4.1.0"
django-debug-toolbar-force = "^0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
